% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cut.R
\name{area_above_ecdf}
\alias{area_above_ecdf}
\title{Area above the eCDF curve}
\usage{
area_above_ecdf(x)
}
\arguments{
\item{x}{A vector of similarity values.}
}
\value{
A numeric value.
}
\description{
Area above the eCDF curve
}
\details{
Denote \code{F(x)} as the eCDF (empirical Cumulative Distribution Function) of the similarity vector \code{x},
this function calculates the area above the eCDF curve, which is \verb{1 - \\int_0^1 F(x)dx}.
}
