\name{cluster_by_apclust}
\alias{cluster_by_apclust}
\title{
Cluster GO similarity matrix by apcluster
}
\description{
Cluster GO similarity matrix by apcluster
}
\usage{
cluster_by_apclust(mat, ...)
}
\arguments{

  \item{mat}{The Go similarity matrix.}
  \item{...}{Other arguments passed to \code{\link[apcluster:apcluster-methods]{apcluster}}.}

}
\examples{
# There is no example
NULL

}
