\name{cluster_by_apcluster}
\alias{cluster_by_apcluster}
\title{
Cluster GO similarity matrix by apcluster
}
\description{
Cluster GO similarity matrix by apcluster
}
\usage{
cluster_by_apcluster(mat, ...)
}
\arguments{

  \item{mat}{The GO similarity matrix.}
  \item{...}{Other arguments passed to \code{\link[apcluster:apcluster-methods]{apcluster}}.}

}
\value{
A vector of cluster labels (in numeric).
}
\examples{
# There is no example
NULL

}
