\name{cluster_by_dynamicTreeCut}
\alias{cluster_by_dynamicTreeCut}
\title{
Cluster GO similarity matrix by dynamicTreeCut
}
\description{
Cluster GO similarity matrix by dynamicTreeCut
}
\usage{
cluster_by_dynamicTreeCut(mat, ...)
}
\arguments{

  \item{mat}{The Go similarity matrix.}
  \item{...}{Other arguments passed to \code{\link[dynamicTreeCut]{cutreeDynamic}}.}

}
\examples{
# There is no example
NULL

}
