\name{cluster_by_kmeans}
\alias{cluster_by_kmeans}
\title{
Cluster GO similarity matrix by k-means clustering
}
\description{
Cluster GO similarity matrix by k-means clustering
}
\usage{
cluster_by_kmeans(mat, ...)
}
\arguments{

  \item{mat}{The Go similarity matrix.}
  \item{...}{Other arguments passed to \code{\link[stats]{kmeans}}.}

}
\examples{
# There is no example
NULL

}
