\name{cluster_by_MCL}
\alias{cluster_by_MCL}
\title{
Cluster similarity matrix by MCL
}
\description{
Cluster similarity matrix by MCL
}
\usage{
cluster_by_MCL(mat, addLoops = TRUE, ...)
}
\arguments{

  \item{mat}{The similarity matrix.}
  \item{addLoops}{Passed to the \code{addLoops} argument in \code{\link[MCL]{mcl}}.}
  \item{...}{Other arguments passed to \code{\link[MCL]{mcl}}.}

}
\value{
A vector of cluster labels (in numeric).
}
\examples{
# There is no example
NULL

}
