\name{cluster_GO}
\alias{cluster_GO}
\title{
Cluster GO terms
}
\description{
Cluster GO terms
}
\usage{
cluster_GO(mat, method = "binary_cut", ...)
}
\arguments{

  \item{mat}{A GO similarity matrix.}
  \item{method}{Method for clustering the matrix.}
  \item{...}{Other arguments passed to the clustering function.}

}
\value{
A numeric vector of cluster labels.
}
\examples{
# There is no example
NULL
}
