\name{[.enrichResult}
\alias{[.enrichResult}
\alias{Extract.enrichResult}
\title{
Subset method of teh enrichResult class
}
\description{
Subset method of teh enrichResult class
}
\usage{
\method{[}{enrichResult}(x, i, j, drop = FALSE)
}
\arguments{

  \item{x}{A \code{enrichResult} object.}
  \item{i}{Row indices.}
  \item{j}{Column indices.}
  \item{drop}{Please ignore.}

}
\examples{
# There is no example
NULL
}
