\name{heightDetails.simple_word_cloud}
\alias{heightDetails.simple_word_cloud}
\title{
Height for simple_word_cloud Grob
}
\description{
Height for simple_word_cloud Grob
}
\usage{
\method{heightDetails}{simple_word_cloud}(x)
}
\arguments{

  \item{x}{The \code{simple_word_cloud} grob returned by \code{\link{simple_word_cloud_grob}}.}

}
\examples{
# There is no example
NULL

}
