\name{heightDetails.word_cloud}
\alias{heightDetails.word_cloud}
\title{
Height for word_cloud Grob
}
\description{
Height for word_cloud Grob
}
\usage{
\method{heightDetails}{word_cloud}(x)
}
\arguments{

  \item{x}{The \code{word_cloud} grob returned by \code{\link{word_cloud_grob}}.}

}
\examples{
# There is no example
NULL

}
