\name{partition_by_kmeanspp}
\alias{partition_by_kmeanspp}
\title{
Partition by kmeans++
}
\description{
Partition by kmeans++
}
\usage{
partition_by_kmeanspp(mat)
}
\arguments{

  \item{mat}{The similarity matrix.}

}
\details{
This function is used to set to the \code{partition_fun} argument in \code{\link{binary_cut}}.
}
\examples{
# There is no example
NULL
}
