\name{partition_by_pam}
\alias{partition_by_pam}
\title{
Partition by PAM
}
\description{
Partition by PAM
}
\usage{
partition_by_pam(mat, k)
}
\arguments{

  \item{mat}{The similarity matrix.}
  \item{k}{Number of clusters.}

}
\details{
The clustering is performed by \code{\link[cluster]{pam}}.
}
\examples{
# There is no example
NULL

}
