\name{plot_binary_cut}
\alias{plot_binary_cut}
\title{
Visualize the process of binary cut
}
\description{
Visualize the process of binary cut
}
\usage{
plot_binary_cut(mat, value_fun = median, cutoff = 0.85,
    partition_fun = partition_by_pam, dend = NULL,
    depth = NULL, show_heatmap_legend = TRUE, ...)
}
\arguments{

  \item{mat}{The similarity matrix.}
  \item{value_fun}{-value_fun}
  \item{cutoff}{-cutoff}
  \item{partition_fun}{-partition_fun}
  \item{dend}{-dend}
  \item{depth}{-depth}
  \item{show_heatmap_legend}{-show_heatmap_legend}
  \item{...}{-...}

}
\examples{
# There is no example
NULL

}
