\name{random_DO}
\alias{random_DO}
\title{
Generate random Disease Ontology (DO) IDs
}
\description{
Generate random Disease Ontology (DO) IDs
}
\usage{
random_DO(n)
}
\arguments{

  \item{n}{Number of DO IDs.}

}
\details{
\code{DO.db} package should be installed.
}
\value{
A vector of DO IDs.
}
\examples{
\donttest{
random_DO(100)
}
}
