\name{remove_clustering_methods}
\alias{remove_clustering_methods}
\title{
Remove clustering methods
}
\description{
Remove clustering methods
}
\usage{
remove_clustering_methods(method)
}
\arguments{

  \item{method}{A vector of method names.}

}
\value{
No value is returned.
}
\examples{
# There is no example
NULL
}
