% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_clusters.R
\name{scale_fontsize}
\alias{scale_fontsize}
\title{Scale font size}
\usage{
scale_fontsize(x, rg = c(1, 30), fs = c(4, 16))
}
\arguments{
\item{x}{A numeric vector.}

\item{rg}{The range.}

\item{fs}{Range of the font size.}
}
\value{
A numeric vector.
}
\description{
Scale font size
}
\details{
It is a linear interpolation.
}
\examples{
x = runif(10, min = 1, max = 20)
# scale x to fontsize 4 to 16.
scale_fontsize(x)
}
