% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{se_opt}
\alias{se_opt}
\title{Global parameters}
\usage{
se_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)
}
\arguments{
\item{...}{Arguments for the parameters, see "details" section.}

\item{RESET}{Whether to reset to default values.}

\item{READ.ONLY}{Please ignore.}

\item{LOCAL}{Please ignore.}

\item{ADD}{Please ignore.}
}
\value{
A \code{GlobalOptionsFun} object.
}
\description{
Global parameters
}
\details{
There are the following global options:
\itemize{
\item \code{verobse}: Whether to print messages.
}
}
