\name{subset_enrichResult}
\alias{subset_enrichResult}
\title{
Subset method of the enrichResult class
}
\description{
Subset method of the enrichResult class
}
\usage{
subset_enrichResult(x, i)
}
\arguments{

  \item{x}{A \code{enrichResult} object from 'clusterProfiler' or other related packages.}
  \item{i}{Row indices.}

}
\value{
Still a \code{enrichResult} object but with the selected subset of rows.
}
\examples{
# There is no example
NULL

}
