\name{term_similarity_from_enrichResult}
\alias{term_similarity_from_enrichResult}
\title{
Similarity between terms in the enrichResult class
}
\description{
Similarity between terms in the enrichResult class
}
\usage{
term_similarity_from_enrichResult(x, ...)
}
\arguments{

  \item{x}{A \code{enrichResult} object.}
  \item{...}{-...}

}
\details{
The object is normally from the clusterProfiler, DOSE, meshes or ReactomePA package.
}
\examples{
# There is no example
NULL

}
