\name{term_similarity_from_Reactome}
\alias{term_similarity_from_Reactome}
\title{
Similarity between Reactome terms
}
\description{
Similarity between Reactome terms
}
\usage{
term_similarity_from_Reactome(term_id, ...)
}
\arguments{

  \item{term_id}{A vector of Reactome IDs.}
  \item{...}{-...}

}
\value{
A symmetric matrix
}
\examples{
# There is no example
NULL

}
