\name{widthDetails.simple_word_cloud}
\alias{widthDetails.simple_word_cloud}
\title{
Width for simple_word_cloud Grob
}
\description{
Width for simple_word_cloud Grob
}
\usage{
\method{widthDetails}{simple_word_cloud}(x)
}
\arguments{

  \item{x}{The \code{simple_word_cloud} grob returned by \code{\link{simple_word_cloud_grob}}.}

}
\examples{
# There is no example
NULL

}
