\name{cluster_by_mclust}
\alias{cluster_by_mclust}
\title{
Cluster GO similarity matrix by mclust
}
\description{
Cluster GO similarity matrix by mclust
}
\usage{
cluster_by_mclust(mat, ...)
}
\arguments{

  \item{mat}{The Go similarity matrix.}
  \item{...}{Other arguments passed to \code{\link[mclust]{Mclust}}.}

}
\examples{
# There is no example
NULL

}
