\name{GO_similarity}
\alias{GO_similarity}
\title{
Calculate GO similarity matrix
}
\description{
Calculate GO similarity matrix
}
\usage{
GO_similarity(go_id, ont, db = 'org.Hs.eg.db', measure = "Rel")
}
\arguments{

  \item{go_id}{A vector of GO ids}
  \item{ont}{Ontology}
  \item{db}{database}
  \item{measure}{Measurement for the GO similarity}

}
\examples{
# There is no example
NULL

}
