\name{partition_by_hclust}
\alias{partition_by_hclust}
\title{
Partition by hclust
}
\description{
Partition by hclust
}
\usage{
partition_by_hclust(mat)
}
\arguments{

  \item{mat}{The similarity matrix.}

}
\details{
The "ward.D2" clusering method was used.

This function is used to set to the \code{partition_fun} argument in \code{\link{binary_cut}}.
}
\examples{
# There is no example
NULL

}
