\name{se_opt}
\alias{se_opt}
\title{
Global parameters
}
\description{
Global parameters
}
\usage{
se_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)
}
\arguments{

  \item{...}{Arguments for the parameters, see "details" section.}
  \item{RESET}{Whether to reset to default values.}
  \item{READ.ONLY}{Please ignore.}
  \item{LOCAL}{Please ignore.}
  \item{ADD}{Please ignore.}

}
\details{
There are the following global options:

\describe{
  \item{\code{verobse}}{Whether to print messages.}
}
}
\examples{
# There is no example
NULL

}
