\name{select_cutoff}
\alias{select_cutoff}
\title{
Select the cutoff for binary cut
}
\description{
Select the cutoff for binary cut
}
\usage{
select_cutoff(mat, cutoff = seq(0.6, 0.9, by = 0.01))
}
\arguments{

  \item{mat}{-mat}
  \item{cutoff}{-cutoff}

}
\examples{
# There is no example
NULL

}
