\name{subset_enrichResult}
\alias{subset_enrichResult}
\title{
Subset method of the enrichResult class
}
\description{
Subset method of the enrichResult class
}
\usage{
subset_enrichResult(x, i)
}
\arguments{

  \item{x}{A \code{enrichResult} object.}
  \item{i}{Row indices.}

}
\examples{
# There is no example
NULL

}
