\name{term_similarity}
\alias{term_similarity}
\title{
Similarity between terms based on the overlap of genes
}
\description{
Similarity between terms based on the overlap of genes
}
\usage{
term_similarity(gl, method = c("kappa", "jaccard"))
}
\arguments{

  \item{gl}{A list of genes that are in the terms.}
  \item{method}{-method}

}
\value{
A symmetric matrix
}
\examples{
# There is no example
NULL

}
