\name{term_similarity_from_KEGG}
\alias{term_similarity_from_KEGG}
\title{
Similarity between KEGG terms
}
\description{
Similarity between KEGG terms
}
\usage{
term_similarity_from_KEGG(term_id)
}
\arguments{

  \item{term_id}{A vector of KEGG IDs.}

}
\value{
A symmetric matrix
}
\examples{
# There is no example
NULL

}
