\name{widthDetails.word_cloud}
\alias{widthDetails.word_cloud}
\title{
Width for word_cloud Grob
}
\description{
Width for word_cloud Grob
}
\usage{
\method{widthDetails}{word_cloud}(x)
}
\arguments{

  \item{x}{The \code{word_cloud} grob returned by \code{\link{word_cloud_grob}}.}

}
\examples{
# There is no example
NULL

}
